+++ web_poet/exceptions/__init__.py
@@ -1,0 +1,2 @@
+from .core import *  # noqa
+from .http import *  # noqa