+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 TBR
 ------------------
 
+* Added support for performing additional requests using
+  ``web_poet.HttpClient``.
+* Introduced ``web_poet.Meta`` to pass arbitrary information
+  inside a Page Object.
 * added a ``PageObjectRegistry`` class which has the  ``handle_urls`` decorator
   to conveniently declare and collect ``OverrideRule``.
 * removed support for Python 3.6